Don't use the {.header.} pragma
Including the header also introduces the compatibility defines which
overwrites the exported proc identifiers.

This reverts commit 8241b9524f5823be75fc4eb3544a4fc269d2c043.

So the fix for https://github.com/nim-lang/Nim/issues/18006 was
to **not** use the {.header.} pragma, so that we don't see `#include
"vpi_compatibility.h"` in the Nim-compiled C.

Also remove the cImport of the whole "veriuser.h". So far, I have
needed it only for tf_dofinish. I am now importc'ing it manually.