Add optional arg allowNilYield to the iterators
The vpiNumArgCheck relies on the iterHandle or argHandle being
nil. While the code in calltf should never have to deal with nil arg
handle or any nil VpiHandle (otherwise we will see crashes at run
time).

So the iterators now have a new arg that allows returning nil iter or
iter element handles; defaults to *not* returning nil handles.

Also, naming is difficult; for now did this renaming and hopefully
that sticks:
- vpiHandles -> vpiHandles2 (because it returns a tuple of 2 elements)
- vpiHandlesWithIndex -> vpiHandles3 (because it returns a tuple of 3
elements)

The iterHandle is now returned as the 2nd element from vpiHandles2 and
as the 3rd element from vpiHandles3; just for aesthetics.